# src/HouseholdLedger.Api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "HouseholdLedger.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/days/{ledgerDate}/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Lists transactions for one ledger date.",
        "parameters": [
          {
            "name": "ledgerDate",
            "in": "path",
            "description": "The route-selected ledger date.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseTransactionResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Creates one transaction for the route-selected ledger date.",
        "parameters": [
          {
            "name": "ledgerDate",
            "in": "path",
            "description": "The route-selected ledger date.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "requestBody": {
          "description": "The request cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseTransactionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseTransactionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseTransactionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/days/{ledgerDate}/transactions/{transactionId}": {
      "put": {
        "tags": [
          "Transactions"
        ],
        "summary": "Corrects one transaction under its route-selected ledger date.",
        "parameters": [
          {
            "name": "ledgerDate",
            "in": "path",
            "description": "The route-selected ledger date.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "transactionId",
            "in": "path",
            "description": "The transaction identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The request cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseTransactionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseTransactionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseTransactionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transactions"
        ],
        "summary": "Removes one transaction under its route-selected ledger date.",
        "parameters": [
          {
            "name": "ledgerDate",
            "in": "path",
            "description": "The route-selected ledger date.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "transactionId",
            "in": "path",
            "description": "The transaction identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateExpenseTransactionRequest": {
        "required": [
          "accountId",
          "amount",
          "classification"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "The owning account identifier.",
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The positive USD amount.",
            "format": "double"
          },
          "classification": {
            "type": "string",
            "description": "The Kakeibo-inspired classification."
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "An optional expense description, at most 200 characters after trimming."
          }
        },
        "description": "Requests creation of one expense for a route-selected date."
      },
...
      "ExpenseTransactionResponse": {
        "required": [
          "id",
          "accountId",
          "accountName",
          "date",
          "amount",
          "classification"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The opaque transaction identifier.",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "description": "The owning account identifier.",
            "format": "uuid"
          },
          "accountName": {
            "type": "string",
            "description": "The owning account name."
          },
          "date": {
            "type": "string",
            "description": "The ledger date.",
            "format": "date"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The positive USD amount.",
            "format": "double"
          },
          "classification": {
            "type": "string",
            "description": "The Kakeibo-inspired classification."
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "The optional normalized expense description."
          }
        },
        "description": "Describes one persisted USD expense."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateExpenseTransactionRequest": {
        "required": [
          "accountId",
          "amount",
          "classification"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "The replacement owning account identifier.",
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The replacement positive USD amount.",
            "format": "double"
          },
          "classification": {
            "type": "string",
            "description": "The replacement Kakeibo-inspired classification."
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "The optional replacement description, at most 200 characters after trimming."
          }
        },
        "description": "Requests correction of one expense without changing its ledger date."
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
...
}